feat: add file upload field component for contract files
feat: create store info card component for displaying store details

feat: implement user confirmation step component for user verification

feat: add user info card component to display registered user information

feat: create user role display component to show user roles with approval status

refactor: streamline user store form by integrating new components and removing redundant code

fix: dynamically import UserStoreForm to improve performance

chore: remove unused SpeedInsights import from layout
