+++ src/app/layout.tsx
@@ -1,6 +1,5 @@
 import { Toaster } from '@/components/ui/sonner';
 import { AntdRegistry } from '@ant-design/nextjs-registry';
-import { SpeedInsights } from '@vercel/speed-insights/next';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import type { Metadata } from 'next';
 import { Roboto } from 'next/font/google';
@@ -33,7 +32,6 @@ export default function RootLayout({
             <Toaster richColors expand visibleToasts={2} />
           </Provider>
         </AntdRegistry>
-        <SpeedInsights />
       </body>
     </html>
   );
